requirements.txt
aniso8601==3.0.2
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-cors-headers==2.4.0
django-filter==2.1.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
idna==2.8
mysqlclient==1.3.14
pkg-resources==0.0.0
promise==2.2.1
pytz==2018.9
requests==2.21.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
urllib3==1.24.1
